# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Pokémon Shiny Encounter Counter — REST API",
        "title": "Encounty API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/update/apply": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Apply a downloaded update",
                "parameters": [
                    {
                        "description": "Download URL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server.UpdateApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
        "/update/check": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Check for available updates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.UpdateInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_server.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_server.UpdateApplyRequest": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string"
                }
            }
        },
...
        "internal_server.UpdateInfo": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "current_version": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "latest_version": {
                    "type": "string"
                }
            }
        },
...
        "internal_server.errResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}